requirements.txt
fastapi~=0.95
orjson~=3.4
pydantic~=1.9
uvicorn~=0.18
uvloop~=0.14
gunicorn~=21.2
SQLAlchemy~=1.4.29
SQLAlchemy-Utils~=0.38
asyncpg~=0.26
databases[postgresql]~=0.6
alembic~=1.11
pytest~=7.4
httpx~=0.24
python-dotenv~=1.0
passlib~=1.7
python-jose~=3.3
bcrypt~=4.0
python-multipart~=0.0.6
redis~=5.0
